move contents of upstream to its mod and fix visibility of methods
