+++ miner-apps/translator/src/lib/sv2/mod.rs
@@ -2,4 +2,4 @@ pub mod channel_manager;
 pub mod upstream;
 
 pub use channel_manager::ChannelManager;
-pub use upstream::upstream::Upstream;
+pub use upstream::Upstream;
